refactor(planning): extract the desktop shell + summary cards (#467)
Second desktop planning slice. Move DesktopPlanningView's dialog/table shell and
summary cards into a new `desktopPlanningCards.tsx`: the DModal dialog wrapper, the
collapsible PlanTable section + its THead, and the AllocationCard budget summary
(with its StackedBar). All module-level and prop-driven — the view still owns the
data + handlers. No behavior or markup change; now-unused imports (useRef,
ChevronDown/ChevronUp, useDialogA11y) drop out of the view.

Parity: the PlanningClient / DesktopPlanningView component tests + the
planning-desktop E2E (which renders the tables + allocation card).

Verified: full unit suite 1261 passed; typecheck 0 errors; lint 0 errors;
git diff --check clean; E2E planning-desktop (2) passed. DesktopPlanningView
936 → 772 lines (1135 → 772 across the desktop split).
